pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.8.2",
    "google-generativeai>=0.8.5",
    "openai>=1.79.0",
    "pandas>=2.2.3",
    "python-dotenv>=1.1.0",
    "selenium>=4.32.0",
    "streamlit>=1.45.1",
    "textract>=1.6.5",
    "trafilatura>=2.0.0",
]
